added express; node backend; concurrently runs client and node
