+++ package.json
@@ -1,16 +1,28 @@
 {
-  "name": "spotify",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-scripts": "1.1.0"
-  },
+  "name": "spotify-mainstream",
+  "version": "1.0.0",
+  "description": "Spotify Mainstream -- an FSE project by Katie & Keenan",
+  "main": "index.js",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "client": "cd client && npm start",
+    "server": "nodemon server.js",
+    "start": "concurrently --kill-others-on-fail \"npm run-script server\" \"npm run-script client\""
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/olin-fse/spotify-mainstream.git"
+  },
+  "author": "Katie & Keenan",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/olin-fse/spotify-mainstream/issues"
+  },
+  "homepage": "https://github.com/olin-fse/spotify-mainstream#readme",
+  "dependencies": {
+    "express": "^4.16.2"
+  },
+  "devDependencies": {
+    "concurrently": "^3.5.1"
   }
 }
